# [pkg/translator/prometheusremotewrite] Improve performance of conversion to Prometheus remote write format